sagemaker/code/requirements.txt
transformers>=4.46.0
torch>=2.1.0
pillow>=9.0.0
accelerate>=0.26.0
huggingface-hub>=0.20.0
